Get template names from tree (#28)
To fully render a wikitext page, a viewer needs to transclude (copy in) the contents of template pages. This requires them to get the template pages and provide them. It'd be convenient if there was a quick way to identify which pages needed to be fetched. Here is that method.