Case insensitive file caching (#54)
## Problem
When caching files on a mac (and windows), filenames are case insensitive, but on linux (and wikipedia itself), they are not. This leads to templates being incorrectly resolved.

## Solution
Cache files with a unique uuid in the filename and a mapping table. This ensures unique mappings even outside of the case sensitivity.